+++ src/runtime_src/xdp/profile/database/static_info/filetypes/aie_control_config_filetype.h
@@ -70,8 +70,8 @@ class AIEControlConfigFiletype : public xdp::aie::BaseFiletypeImpl {
                           const std::string& metricStr = "channels",
                           int16_t channelId = -1,
                           bool useColumn = false, 
-                          uint32_t minCol = 0, 
-                          uint32_t maxCol = 0) override; 
+                          uint8_t minCol = 0, 
+                          uint8_t maxCol = 0) override; 
 
         std::vector<tile_type>
         getMemoryTiles(const std::string& graphName,
